Add entry point of the project.
This module is the module that will link all the jobs. The goal is to parse an
input file and read all its lines. Each line will be cut into words, and the
words will be sanitized be fore being analyzed by enchant module. In case of
error, it will be sent to the module 'prompt' in order to suggest a correction.